Add files via upload
This file is used to store the latest update of patient record after interrupting the terminal and from which we can reload the latest update imposed to database when executing the .py file in terminal. This file is not displayed in VSCode because it's either binary or uses unsupported text encoding and should be in the same exact folder with .py file.